[CSS Modules] Fix crash in GetFetchedStyleSheetFromModuleMap
Document::DOMWindow is not always available, for example when parsing
via Document.parseHTMLUnsafe. This change adds an early return to
prevent crashing. This means declarative modules won't work via
Document.parseHTMLUnsafe, but that matches the behavior of imperative
modules. I added a TODO to follow up with the WHATWG about this.

A test was added that crashes without the fix.

Bug: 448174611
Change-Id: I471d397fd730385d16eaef231c3cb8c369230597