Fix corner-shape shadow & outline rendering
As per resolution in
https://github.com/w3c/csswg-drafts/issues/13037#issuecomment-3582530992

Shadows should be border-aligned, like border inset,
with external "miters" that are clipped to the outer edge.

The miter is computed by extending the start and end tangents of the
curve until they reach the outer edge.

The tangent is the line between the start point, and the "control
point": a point that would be the quadratic control point if this
was a normal quadratic curve with the same half-corner.

As part of this work, refactoring the rendering tests to be both
more precise and closer to something that can be reasoned about being
close to the spec.

See spec PR: https://github.com/w3c/csswg-drafts/pull/13173
The PR and the WPT rendering code were developed side by side to
ensure that the tests assert the spec.

Since the spec defines that curves are approximate, and ref tests can
fail on antialiasing and curve inaccuracies, the corners are also
stroked with a blue 3px line on both the actual and ref versions, to
avoid test failures in these occasions.

(This change is guarded by the BorderRadiusCorrectionCoverageFactor,
as without it shadow rects wouldn't have an origin rect, so the
changed code paths is never reached).

Bug: 463996869
Change-Id: I7f586440f47757416b8eb9a2002629cd29e0174a